+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased
 
 - Fix a bug that breaks hidden fields of type reference
+- Add dynamic method "field" to model mixin
 
 # 0.0.5